+++ Source/Bugsnag.m
@@ -30,6 +30,7 @@
 #import "BugsnagClient.h"
 #import "BugsnagKeys.h"
 #import "BugsnagPlugin.h"
+#import "BugsnagHandledState.h"
 
 static BugsnagClient *bsg_g_bugsnag_client = NULL;
 
@@ -47,8 +48,16 @@ static BugsnagClient *bsg_g_bugsnag_client = NULL;
 - (NSDictionary *)BSG_mergedInto:(NSDictionary *)dest;
 @end
 
+@interface BugsnagEvent ()
+@property(readwrite) NSUInteger depth;
+@end
+
 @interface BugsnagClient ()
 - (void)startListeningForStateChangeNotification:(NSString *_Nonnull)notificationName;
+@end
+
+@interface BugsnagMetadata ()
+- (NSDictionary *_Nonnull)toDictionary;
 @end
 
 @implementation Bugsnag