+++ Tests/BugsnagEventTests.m
@@ -21,6 +21,16 @@
 + (BugsnagConfiguration *)configuration;
 @end
 
+@interface BugsnagEvent ()
+- (NSDictionary *_Nonnull)toJson;
+- (BOOL)shouldBeSent;
+@property(readwrite) NSUInteger depth;
+@end
+
+@interface BugsnagMetadata ()
+- (NSDictionary *_Nonnull)toDictionary;
+@end
+
 @interface BugsnagEventTests : BugsnagBaseUnitTest
 @end